basic form and value retrieval working
