+++ src/index.ts
@@ -1,3 +1,3 @@
-import { Form } from "./core";
-import { useInput } from "./react_adapters";
-export { Form };
+import { createForm } from "./core";
+// import { useInput } from "./react_adapters";
+export { createForm };
